*: Only use multiarch if necessary

+++ scripts/docker-build-push.sh
@@ -43,20 +43,21 @@ esac
 
 PUSH_CONTEXT=${REPOSITORY}/${IMAGE}:${RC:-${TAG}${BRANCH}}
 
-docker context create multiarch 2> /dev/null || true
-
-
-if [ "$(uname)" = "Linux" ]; then
-  docker run --rm --privileged multiarch/qemu-user-static --reset -p yes
-fi
-
-echo ::group::Use buildx-multiarch
-docker buildx use buildx-multiarch ||
-  docker buildx create --driver docker-container --use multiarch --name buildx-multiarch
-echo ::end"group"::
-
 if [ -n "$PLATFORMS" ]; then
   PLATFORM_ARGS="--platform $PLATFORMS"
+
+  if [[ "$PLATFORMS" == *,* ]]; then
+    docker context create multiarch 2> /dev/null || true
+
+    if [ "$(uname)" = "Linux" ]; then
+      docker run --rm --privileged multiarch/qemu-user-static --reset -p yes
+    fi
+
+    echo ::group::Use buildx-multiarch
+    docker buildx use buildx-multiarch ||
+      docker buildx create --driver docker-container --use multiarch --name buildx-multiarch
+    echo ::end"group"::
+  fi
 fi
 
 if [ -n "$BUILD_ARGS" ]; then
